double to float for MVA

+++ embedded/THStack_plot_embedded_FBDT.c
@@ -279,7 +279,7 @@ void LetsFillJpsi(const char* dirname, std::vector<std::string> variable_names, 
     13: hhISR
     */
 
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
     double Upsilon_ID = -1;
     double Bsig_ID = -1;
     double temp_N_bin_PID[4][N_PID_syst] = { 0.0 }; // K-true, K-mis, pi-true, pi-miss
@@ -476,7 +476,7 @@ void LetsFillJpsi(const char* dirname, std::vector<std::string> variable_names, 
 
 void LetsFillJpsi(const char* dirname, std::vector<std::string> variable_names, std::vector<std::string> branch_names, std::vector<double> variable_values[Nvar_num], std::vector<double>* weights, double weight) {
 
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
     double Upsilon_ID = -1;
     double Bsig_ID = -1;
 
@@ -539,7 +539,7 @@ void LetsFillJpsi_ri(const char* dirname, std::vector<std::string> variable_name
     13: hhISR
     */
 
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
     double Upsilon_ID = -1;
     double Bsig_ID = -1;
     double temp_N_bin_PID[4][N_PID_syst] = { 0.0 }; // K-true, K-mis, pi-true, pi-miss
@@ -775,7 +775,7 @@ void LetsFillJpsi_ri(const char* dirname, std::vector<std::string> variable_name
 }
 
 void LetsFill(const char* dirname, std::vector<std::string> variable_names, std::vector<std::string> branch_names, std::vector<double> variable_values[Nvar_num]) {
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
 
     std::vector<string> names;
     load_files(dirname, &names);
@@ -814,7 +814,7 @@ void LetsFill(const char* dirname, std::vector<std::string> variable_names, std:
 }
 
 void LetsFill(const char* dirname, std::vector<std::string> variable_names, std::vector<std::string> branch_names, std::vector<double> variable_values[Nvar_num], const char* included_string) {
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
 
     std::vector<string> names;
     load_files(dirname, &names, included_string);
@@ -870,7 +870,7 @@ void LetsFillJpsi_ri_correction(const char* dirname, std::vector<std::string> va
     13: hhISR
     */
 
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
     double Upsilon_ID = -1;
     double Bsig_ID = -1;
     double temp_N_bin_PID[4][N_PID_syst] = { 0.0 }; // K-true, K-mis, pi-true, pi-miss
@@ -1128,7 +1128,7 @@ void NevtCount_ri(const char* dirname, std::string SampleName, Nevt* nevt) {
     14: signal (B-->Xs J/psi)
     */
 
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
     double Upsilon_ID = -1;
     double Bsig_ID = -1;
     double temp_N_bin_PID[4][N_PID_syst] = { 0.0 }; // K-true, K-mis, pi-true, pi-miss
