+++ side_band/THStack_plot_sideband_FBDT.c
@@ -279,7 +279,7 @@ void LetsFillSideBand(const char* dirname, std::vector<std::string> variable_nam
     13: hhISR
     */
 
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
     double Upsilon_ID = -1;
     double Bsig_ID = -1;
     double temp_N_bin_PID[4][N_PID_syst] = { 0.0 }; // K-true, K-mis, pi-true, pi-miss
@@ -492,7 +492,7 @@ void LetsFillSideBand_ri(const char* dirname, std::vector<std::string> variable_
     13: hhISR
     */
 
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
     double Upsilon_ID = -1;
     double Bsig_ID = -1;
     double temp_N_bin_PID[4][N_PID_syst] = { 0.0 }; // K-true, K-mis, pi-true, pi-miss
@@ -694,7 +694,7 @@ void LetsFillSideBand_ri(const char* dirname, std::vector<std::string> variable_
 }
 
 void LetsFill(const char* dirname, std::vector<std::string> variable_names, std::vector<std::string> branch_names, std::vector<double> variable_values[Nvar_num]) {
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
 
     std::vector<string> names;
     load_files(dirname, &names);
@@ -733,7 +733,7 @@ void LetsFill(const char* dirname, std::vector<std::string> variable_names, std:
 }
 
 void LetsFill(const char* dirname, std::vector<std::string> variable_names, std::vector<std::string> branch_names, std::vector<double> variable_values[Nvar_num], const char* included_string) {
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
 
     std::vector<string> names;
     load_files(dirname, &names, included_string);
@@ -789,7 +789,7 @@ void LetsFillSideBand_ri_correction(const char* dirname, std::vector<std::string
     13: hhISR
     */
 
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
     double Upsilon_ID = -1;
     double Bsig_ID = -1;
     double temp_N_bin_PID[4][N_PID_syst] = { 0.0 }; // K-true, K-mis, pi-true, pi-miss
@@ -1022,7 +1022,7 @@ void NevtCount_ri(const char* dirname, std::string SampleName, Nevt* nevt) {
     13: hhISR
     */
 
-    double var[Nvar_num] = { 0.0 };
+    float var[Nvar_num] = { 0.0 };
     double Upsilon_ID = -1;
     double Bsig_ID = -1;
     double temp_N_bin_PID[4][N_PID_syst] = { 0.0 }; // K-true, K-mis, pi-true, pi-miss
